+++ src/components/StatusBar/index.ts
@@ -1,0 +1,3 @@
+import { StatusBar } from "./StatusBar";
+
+export default StatusBar;
